+++ src/adjusters/Tax.php
@@ -139,7 +139,7 @@ class Tax extends Component implements AdjusterInterface
             // before we return false (no taxes) remove the tax if it was included in the taxable amount.
             if ($taxRate->include) {
                 // Is this an order level tax rate?
-                if (in_array($taxRate->taxable, [TaxRateRecord::TAXABLE_ORDER_TOTAL_PRICE, TaxRateRecord::TAXABLE_ORDER_TOTAL_SHIPPING], false)) {
+                if (in_array($taxRate->taxable, TaxRateRecord::ORDER_TAXABALES, false)) {
                     $orderTaxableAmount = 0;
 
                     if ($taxRate->taxable === TaxRateRecord::TAXABLE_ORDER_TOTAL_PRICE) {
@@ -196,7 +196,7 @@ class Tax extends Component implements AdjusterInterface
         }
 
         // Is this an order level tax rate?
-        if (in_array($taxRate->taxable, [TaxRateRecord::TAXABLE_ORDER_TOTAL_PRICE, TaxRateRecord::TAXABLE_ORDER_TOTAL_SHIPPING], false)) {
+        if (in_array($taxRate->taxable, TaxRateRecord::ORDER_TAXABALES, false)) {
 
             $allItemsTaxFree = true;
             foreach ($this->_order->getLineItems() as $item) {