+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed a bug where the "Download PDF" button wouldn't show in Orders. ([#962](https://github.com/craftcms/commerce/issues/962))
 - Fixed a bug where save shortcut didn't work in "General Settings".
 - Fixed a bug where save shortcut didn't work in "Store Location" settings.
+- Fixed a bug that forced the user to choose a "Tax Category" for order "Taxable Subjects". ([#538](https://github.com/craftcms/commerce/issues/538))
 
 ## 2.1.12.1 - 2019-08-23